pass along rerun mode
