+++ sematic/api_client.py
@@ -26,6 +26,7 @@ from sematic.db.models.job import Job
 from sematic.db.models.resolution import Resolution
 from sematic.db.models.run import Run
 from sematic.db.models.user import User
+from sematic.graph import RerunMode
 from sematic.metrics.metric_point import MetricPoint
 from sematic.plugins.abstract_external_resource import AbstractExternalResource
 from sematic.utils.retry import retry, retry_call
@@ -296,6 +297,7 @@ def schedule_resolution(
     resolution_id: str,
     max_parallelism: Optional[int] = None,
     rerun_from: Optional[str] = None,
+    rerun_mode: Optional[RerunMode] = None,
 ) -> Resolution:
     """Ask the server to start a detached resolution execution."""
     payload: Dict[str, Any] = {}
@@ -305,6 +307,9 @@ def schedule_resolution(
 
     if rerun_from is not None:
         payload["rerun_from"] = rerun_from
+
+    if rerun_mode is not None:
+        payload["rerun_mode"] = rerun_mode
 
     response = _post(f"/resolutions/{resolution_id}/schedule", json_payload=payload)
     return Resolution.from_json_encodable(response["content"])
