+++ sematic/resolvers/BUILD
@@ -74,11 +74,11 @@ sematic_py_lib(
         "//sematic:abstract_future",
         "//sematic:api_client",
         "//sematic:container_images",
-        "//sematic:graph",
         "//sematic/db/models:artifact",
         "//sematic/db/models:edge",
         "//sematic/db/models:resolution",
         "//sematic/db/models:run",
+        "//sematic:graph",
         "//sematic/plugins:abstract_external_resource",
         "//sematic/utils:exceptions",
         "//sematic/utils:memoized_property",
@@ -125,6 +125,7 @@ sematic_py_lib(
         "//sematic:function",
         "//sematic:future",
         "//sematic:future_context",
+        "//sematic:graph",
         "//sematic:log_reader",
         "//sematic:versions",
         "//sematic/config:user_settings",
@@ -132,6 +133,7 @@ sematic_py_lib(
         "//sematic/db/models:edge",
         "//sematic/db/models:factories",
         "//sematic/resolvers:cloud_resolver",
+        "//sematic/resolvers:local_resolver",
         "//sematic/scheduling:job_details",
         "//sematic/utils:exceptions",
     ],
